Release v0.0.1-4431

+++ feature/b2b/bgp/ebgp_route_prefix_test.go
@@ -90,7 +90,7 @@ func ebgpRoutePrefixConfig(api *otg.OtgApi, tc map[string]interface{}) gosnappi.
 		SetName("dtxIp").
 		SetAddress(tc["txIp"].(string)).
 		SetGateway(tc["txGateway"].(string)).
-		SetPrefix(tc["txPrefix"].(int32))
+		SetPrefix(tc["txPrefix"].(uint32))
 
 	dtxBgp := dtx.Bgp().
 		SetRouterId(tc["txIp"].(string))
@@ -102,7 +102,7 @@ func ebgpRoutePrefixConfig(api *otg.OtgApi, tc map[string]interface{}) gosnappi.
 	dtxBgpv4Peer := dtxBgpv4.
 		Peers().
 		Add().
-		SetAsNumber(tc["txAs"].(int32)).
+		SetAsNumber(tc["txAs"].(uint32)).
 		SetAsType(gosnappi.BgpV4PeerAsType.EBGP).
 		SetPeerAddress(tc["txGateway"].(string)).
 		SetName("dtxBgpv4Peer")
@@ -120,7 +120,7 @@ func ebgpRoutePrefixConfig(api *otg.OtgApi, tc map[string]interface{}) gosnappi.
 	dtxBgpv4PeerRrV4.Addresses().Add().
 		SetAddress(tc["txAdvRouteV4"].(string)).
 		SetPrefix(32).
-		SetCount(tc["txRouteCount"].(int32)).
+		SetCount(tc["txRouteCount"].(uint32)).
 		SetStep(1)
 
 	dtxBgpv4PeerRrV4.Advanced().
@@ -136,7 +136,7 @@ func ebgpRoutePrefixConfig(api *otg.OtgApi, tc map[string]interface{}) gosnappi.
 		SetAsSetMode(gosnappi.BgpAsPathAsSetMode.INCLUDE_AS_SET)
 
 	dtxBgpv4PeerRrV4AsPath.Segments().Add().
-		SetAsNumbers([]int64{1112, 1113}).
+		SetAsNumbers([]uint32{1112, 1113}).
 		SetType(gosnappi.BgpAsPathSegmentType.AS_SEQ)
 
 	dtxBgpv4PeerRrV6 := dtxBgpv4Peer.
@@ -150,7 +150,7 @@ func ebgpRoutePrefixConfig(api *otg.OtgApi, tc map[string]interface{}) gosnappi.
 	dtxBgpv4PeerRrV6.Addresses().Add().
 		SetAddress(tc["txAdvRouteV6"].(string)).
 		SetPrefix(128).
-		SetCount(tc["txRouteCount"].(int32)).
+		SetCount(tc["txRouteCount"].(uint32)).
 		SetStep(1)
 
 	dtxBgpv4PeerRrV6.Advanced().
@@ -166,7 +166,7 @@ func ebgpRoutePrefixConfig(api *otg.OtgApi, tc map[string]interface{}) gosnappi.
 		SetAsSetMode(gosnappi.BgpAsPathAsSetMode.INCLUDE_AS_SET)
 
 	dtxBgpv4PeerRrV6AsPath.Segments().Add().
-		SetAsNumbers([]int64{1112, 1113}).
+		SetAsNumbers([]uint32{1112, 1113}).
 		SetType(gosnappi.BgpAsPathSegmentType.AS_SEQ)
 
 	drxEth := drx.Ethernets().
@@ -183,7 +183,7 @@ func ebgpRoutePrefixConfig(api *otg.OtgApi, tc map[string]interface{}) gosnappi.
 		SetName("drxIp").
 		SetAddress(tc["rxIp"].(string)).
 		SetGateway(tc["rxGateway"].(string)).
-		SetPrefix(tc["rxPrefix"].(int32))
+		SetPrefix(tc["rxPrefix"].(uint32))
 
 	drxBgp := drx.Bgp().
 		SetRouterId(tc["rxIp"].(string))
@@ -195,7 +195,7 @@ func ebgpRoutePrefixConfig(api *otg.OtgApi, tc map[string]interface{}) gosnappi.
 	drxBgpv4Peer := drxBgpv4.
 		Peers().
 		Add().
-		SetAsNumber(tc["rxAs"].(int32)).
+		SetAsNumber(tc["rxAs"].(uint32)).
 		SetAsType(gosnappi.BgpV4PeerAsType.EBGP).
 		SetPeerAddress(tc["rxGateway"].(string)).
 		SetName("drxBgpv4Peer")
@@ -213,7 +213,7 @@ func ebgpRoutePrefixConfig(api *otg.OtgApi, tc map[string]interface{}) gosnappi.
 	drxBgpv4PeerRrV4.Addresses().Add().
 		SetAddress(tc["rxAdvRouteV4"].(string)).
 		SetPrefix(32).
-		SetCount(tc["rxRouteCount"].(int32)).
+		SetCount(tc["rxRouteCount"].(uint32)).
 		SetStep(1)
 
 	drxBgpv4PeerRrV4.Advanced().
@@ -229,7 +229,7 @@ func ebgpRoutePrefixConfig(api *otg.OtgApi, tc map[string]interface{}) gosnappi.
 		SetAsSetMode(gosnappi.BgpAsPathAsSetMode.INCLUDE_AS_SET)
 
 	drxBgpv4PeerRrV4AsPath.Segments().Add().
-		SetAsNumbers([]int64{4444}).
+		SetAsNumbers([]uint32{4444}).
 		SetType(gosnappi.BgpAsPathSegmentType.AS_SEQ)
 
 	drxBgpv4PeerRrV6 := drxBgpv4Peer.
@@ -243,7 +243,7 @@ func ebgpRoutePrefixConfig(api *otg.OtgApi, tc map[string]interface{}) gosnappi.
 	drxBgpv4PeerRrV6.Addresses().Add().
 		SetAddress(tc["rxAdvRouteV6"].(string)).
 		SetPrefix(128).
-		SetCount(tc["rxRouteCount"].(int32)).
+		SetCount(tc["rxRouteCount"].(uint32)).
 		SetStep(1)
 
 	drxBgpv4PeerRrV6.Advanced().
@@ -259,14 +259,14 @@ func ebgpRoutePrefixConfig(api *otg.OtgApi, tc map[string]interface{}) gosnappi.
 		SetAsSetMode(gosnappi.BgpAsPathAsSetMode.INCLUDE_AS_SET)
 
 	drxBgpv4PeerRrV6AsPath.Segments().Add().
-		SetAsNumbers([]int64{4444}).
+		SetAsNumbers([]uint32{4444}).
 		SetType(gosnappi.BgpAsPathSegmentType.AS_SEQ)
 
 	for i := 1; i <= 4; i++ {
 		flow := c.Flows().Add()
-		flow.Duration().FixedPackets().SetPackets(tc["pktCount"].(int32))
-		flow.Rate().SetPps(tc["pktRate"].(int64))
-		flow.Size().SetFixed(tc["pktSize"].(int32))
+		flow.Duration().FixedPackets().SetPackets(tc["pktCount"].(uint32))
+		flow.Rate().SetPps(tc["pktRate"].(uint64))
+		flow.Size().SetFixed(tc["pktSize"].(uint32))
 		flow.Metrics().SetEnable(true)
 	}
 
@@ -345,8 +345,8 @@ func ebgpRoutePrefixConfig(api *otg.OtgApi, tc map[string]interface{}) gosnappi.
 func ebgpRoutePrefixBgpMetricsOk(api *otg.OtgApi, tc map[string]interface{}) bool {
 	for _, m := range api.GetBgpv4Metrics() {
 		if m.SessionState() == gosnappi.Bgpv4MetricSessionState.DOWN ||
-			m.RoutesAdvertised() != 2*tc["txRouteCount"].(int32) ||
-			m.RoutesReceived() != 2*tc["rxRouteCount"].(int32) {
+			m.RoutesAdvertised() != 2*tc["txRouteCount"].(uint64) ||
+			m.RoutesReceived() != 2*tc["rxRouteCount"].(uint64) {
 			return false
 		}
 	}
@@ -375,7 +375,7 @@ func ebgpRoutePrefixBgpPrefixesOk(api *otg.OtgApi, tc map[string]interface{}) bo
 }
 
 func ebgpRoutePrefixFlowMetricsOk(api *otg.OtgApi, tc map[string]interface{}) bool {
-	pktCount := int64(tc["pktCount"].(int32))
+	pktCount := uint64(tc["pktCount"].(int32))
 
 	for _, m := range api.GetFlowMetrics() {
 		if m.Transmit() != gosnappi.FlowMetricTransmit.STOPPED ||
